% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_A.R
\name{best_A}
\alias{best_A}
\title{Find the Best Quantile for A quantile cutoff in \code{filter_params} {HiCcompare}}
\usage{
best_A(hic.table, SD = 2, numChanges = 35, FC = 3, alpha = 0.05, Plot = FALSE)
}
\arguments{
\item{hic.table}{A hic.table object}

\item{SD}{Numeric. The standard deviation of the fuzzing used to produce a Hi-C matrix from your data with few true differences.}

\item{numChanges}{Integer. The number of changes to add into the Hi-C matrix created. This should be proportional to the resolution of the data. High resolution data should use more changes i.e. 1MB resolution - 300 changes, 100KB resolution - 1000 changes, etc.}

\item{FC}{Numeric. The fold change of the changes added to the Hi-C matrix.}

\item{alpha}{Numeric. Significance level for adjusting p-values (default is 0.05).}

\item{Plot}{Logical. If TRUE, plots will be generated during the processing.}
}
\value{
A data frame containing the best quantile for A, the corresponding best A value,
and performance metrics (MCC, TPR, FPR).
}
\description{
This function identifies the best quantile for the parameter A quantile cutoff in \code{filter_params} {HiCcompare} by evaluating
different quantile levels based on performance metrics such as
Matthews Correlation Coefficient (MCC), True Positive Rate (TPR), and
False Positive Rate (FPR).
}
\details{
The function randomizes interaction frequencies, calculates mean interaction
frequencies, and performs a series of evaluations to determine the best quantile.
The results include the quantile value, best A value, and performance metrics for
each evaluated quantile.
}
\examples{
# Assuming 'hic.table' is a valid Hi-C interaction frequency table
best_result <- best_quantile_A(hic.table, SD = 2, numChanges = 35, FC = 3, alpha = 0.05)
print(best_result)

}
