% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_example_MGfolder.R
\name{load_example_MGfolder}
\alias{load_example_MGfolder}
\title{Load MG Datasets Folder Example}
\usage{
load_example_MGfolder(save_dir = NULL)
}
\arguments{
\item{save_dir}{A character string specifying the directory where the 'MGs_example'
folder will be created. If NULL (default), the function uses the current working directory.}
}
\value{
Saves the datasets as '.txt' files in the 'MGs_example' folder.
}
\description{
This function loads example of MG datasets and saves them in a newly created folder named 'MGs_example'
within the specified directory. This is the example data folder to be used as example for \code{scHiC_table} function. If no directory is specified, it defaults to
the current working directory.
}
\examples{
load_example_MGfolder() # Saves in 'MGs_example' in the current working directory
load_example_MGfolder("/path/to/save/directory") # Saves in specified directory

}
