% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_example_ODCfolder.R
\name{load_example_ODCfolder}
\alias{load_example_ODCfolder}
\title{Load ODC Datasets Folder Example}
\usage{
load_example_ODCfolder(save_dir = NULL)
}
\arguments{
\item{save_dir}{A character string specifying the directory where the 'ODCs_example'
folder will be created. If NULL (default), the function uses the current working directory.}
}
\value{
Saves the datasets as '.txt' files in the 'ODCs_example' folder.
}
\description{
This function loads example of ODC datasets and saves them in a newly created folder named 'ODCs_example'
within the specified directory. This is the example data folder to be used as example for \code{scHiC_table} function. If no directory is specified, it defaults to
the current working directory.
}
\examples{
load_example_ODCfolder() # Saves in 'ODCs_example' in the current working directory
load_example_ODCfolder("/path/to/save/directory") # Saves in specified directory

}
