% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HiCmatrix_heatmap.R
\name{plot_HiCmatrix_heatmap}
\alias{plot_HiCmatrix_heatmap}
\title{Plot Hi-C Interaction Matrix Heatmap}
\usage{
plot_HiCmatrix_heatmap(
  scHiC.sparse,
  zlim = NULL,
  color_low = "white",
  color_high = "red",
  main = NULL,
  figure_name = NULL
)
}
\arguments{
\item{scHiC.sparse}{A modified sparse matrix of Hi-C interaction frequencies in format (chr1, start1, chr2, start2, IF).}

\item{zlim}{A numeric vector of length 2 specifying the limits of the color scale. If the zlim is not specified, the zlim vector include minimum and maximum values of the matrix.}

\item{color_low}{A character specifying the color for the lowest values. Default is "white".}

\item{color_high}{A character specifying the color for the highest values. Default is "red".}

\item{main}{A character string for the main title of the plot. Default is NULL.}

\item{figure_name}{A character string for additional figure labeling. Default is NULL.}
}
\value{
A heatmap plot visualizing the Hi-C interaction matrix.
}
\description{
This function generates a heatmap of Hi-C interaction frequencies from a given sparse matrix,
allowing users to visualize either original or imputed Hi-C data.
}
\examples{
# Create a 36x36 matrix with random integer values between 0 and 9
matrix_36x36 <- matrix(sample(0:9, 36*36, replace = TRUE), nrow = 36, ncol = 36)
diag(matrix_36x36) <- diag(matrix_36x36) * 6

# Transform full matrix into sparse
library(HiCcompare)
sparse = full2sparse(matrix_36x36)

# Call the function with this sparse matrix to generate the heatmap
plot_imputation_heatmap(
 scHiC.sparse = sparse, 
 zlim = c(0, 7),              # Log scale color limits
 color_low = "white",          # Color for low values
 color_high = "red",           # Color for high values
 main = "Single-Cell Hi-C Heatmap",  # Title of the plot
 figure_name = "Example Heatmap"     # Subtitle for the plot
)

}
