% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_imputation_heatmap.R
\name{plot_imputation_heatmap}
\alias{plot_imputation_heatmap}
\title{Plot Hi-C Interaction Matrix Heatmap}
\usage{
plot_imputation_heatmap(
  scHiC.sparse,
  zlim = c(0, 7),
  color_low = "white",
  color_high = "red",
  main = NULL,
  figure_name = NULL
)
}
\arguments{
\item{scHiC.sparse}{A sparse matrix of Hi-C interaction frequencies in sparse format.}

\item{zlim}{A numeric vector of length 2 specifying the limits of the color scale. Default is c(0, 7).}

\item{color_low}{A character specifying the color for the lowest values. Default is "white".}

\item{color_high}{A character specifying the color for the highest values. Default is "red".}

\item{main}{A character string for the main title of the plot. Default is NULL.}

\item{figure_name}{A character string for additional figure labeling. Default is NULL.}
}
\value{
A heatmap plot visualizing the Hi-C interaction matrix.
}
\description{
This function generates a heatmap of Hi-C interaction frequencies from a given sparse matrix,
allowing users to visualize either original or imputed Hi-C data.
}
