% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_imputed_distance_diagnostic.R
\name{plot_imputed_distance_diagnostic}
\alias{plot_imputed_distance_diagnostic}
\title{Plot Imputed Distance Diagnostic}
\usage{
plot_imputed_distance_diagnostic(org_sc_data, imp_sc_data, D)
}
\arguments{
\item{org_sc_data}{A data frame containing the original single-cell Hi-C data in scHiC Table object.
The data frame should include columns for \code{region1},\code{region2},\code{Cell}, \code{Chr}, and \code{IF_i} for i single cells.}

\item{imp_sc_data}{A data frame containing the imputed single-cell Hi-C datain scHiC Table object.
The data frame should include columns for \code{region1},\code{region2},\code{Cell}, \code{Chr}, and \code{IF_i} for i single cells.}

\item{D}{An integer specifying the genomic distance for which to plot the density
curves of interaction frequencies.}
}
\value{
A ggplot2 object representing the density plot comparing original
and imputed interaction frequencies.
}
\description{
This function plots the density curves of interaction frequencies
from original and imputed single-cell Hi-C data for a specified distance.
}
\examples{
# Assuming scHiC.table_ODC and imp.cc.table.1 are your data frames:
plot_imputed_distance_diagnostic(scHiC.table_ODC, imp.cc.table.1, D = 3)

}
