% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHiC_table.R
\name{read_files}
\alias{read_files}
\title{Read Files for scHi-C Dataset}
\usage{
read_files(file.path, cell, position.dataset, type = "txt")
}
\arguments{
\item{file.path}{The directory path where the data files are stored.}

\item{cell}{A character string specifying a cell type (eg., 'MG', etc).}

\item{position.dataset}{A vector of indices specifying the file positions to read from the directory.}

\item{type}{The file type, either 'txt' or 'cool'. Default is 'txt'. Each 'txt; file should be in the format of a sparse upper triangular Hi-C matrix with five columns:
chromosome 1 (chr1), start position 1 (start1), chromosome 2 (chr2), start position 2 (start2), and interaction frequency (IF).
If the 'cool' type is selected, the .cool files are HDF5 containers that store Hi-C data, which will be read using \code{cooler2bedpe()} from the \code{HiCcompare} package.}
}
\value{
A list of datasets, where each element corresponds to a dataset from the selected files. If 'txt' type is selected, each element contains a data frame of a sparse upper triangular Hi-C matrix.
If 'cool' is selected, each element contains a list with two items:
\itemize{
\item "cis": Contains the intra-chromosomal contact matrices, one per chromosome.
\item "trans": Contains the inter-chromosomal contact matrix.
}
}
\description{
This function reads single-cell Hi-C data from a specified file path.
It supports two file formats: 'txt' and 'cool'. For 'txt', it reads tab-delimited files and assumes the format contains five columns:
chromosome 1 (chr1), start position 1 (start1), chromosome 2 (chr2), start position 2 (start2), and interaction frequency (IF).
For 'cool', it uses the \code{HiCcompare} package to transform cooler files to BEDPE format.
}
\details{
This function input single cell Hi-C data in '.txt' or 'cool' format. Each 'txt; file should be in the form of a sparse upper triangular Hi-C matrix with five columns:
chromosome 1 (chr1), start genomic position 1 (start1), chromosome 2 (chr2), start genomic position 2 (start2), and interaction frequency (IF), with format of tab-separated columns without row names, column names, or quotes around character strings.
If the 'cool' type is selected, the .cool files are HDF5 containers that store Hi-C data, which will be read using \code{cooler2bedpe()} from the \code{HiCcompare} package.
}
\examples{
\dontrun{
datasets <- read_files("scHiCcompare/Example/Data/MG", "NSN", c(1, 2, 3), type = "txt")
}
}
