% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHiC.table_ODC_chr22.R
\docType{data}
\name{scHiC.table_ODC_chr22}
\alias{scHiC.table_ODC_chr22}
\title{scHi-C table from oligodendrocyte (ODC) cell type - chromosome 22 at 1 MB resolution}
\format{
A dataframe with 335 rows and 5 columns:
\describe{
\item{cell}{Name of the target cell type }
\item{chr}{The interacting region starting cordination chromosome}
\item{start1}{The first interacting region starting cordination}
\item{start2}{The second interacting region starting cordination}
\item{IF_i}{The interaction frequency value for single cells, where i ranges from 1 to 10}
}
}
\source{
Single-cells data of ODC cell type download by download_schic() by \code{Bandnorm}. See their website at
\url{https://sshen82.github.io/BandNorm/articles/BandNorm-tutorial.html#download-existing-single-cell-hi-c-data}
}
\usage{
scHiC.table_ODC_chr22
}
\value{
A data frame
}
\description{
A table gather 10 single cells Hi-C data modified in sparse upper triangular matrix containing
the cell's interacting regions cordination (cell, chr, start1, start2)
and the corresponding Interaction Frequencies for each single cell (IF_1, IF2,... IF10).
}
\keyword{datasets}
