% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScHiCcompare.R
\name{ScHiCcompare}
\alias{ScHiCcompare}
\title{ScHiCcompare: Differential Analysis of Single-Cell Hi-C Data}
\usage{
ScHiCcompare(
  file.path.1,
  file.path.2,
  imputation = "RF",
  normalization = "Loess",
  differential.detect = "MD.cluster",
  select.chromosome,
  main.Distances = 1:10000000,
  save.output.path = NULL,
  Plot = T,
  Plot.normalize = F,
  pool.style = "progressive",
  n.imputation = 5,
  maxit = 1,
  outlier.rm = TRUE,
  missPerc.threshold = 95,
  A.min = NULL,
  fprControl.logfc = 1,
  alpha = 0.05,
  BP_param = bpparam()
)
}
\arguments{
\item{file.path.1}{Character string specifying the directory containing scHi-C data for the first condition (first cell-type group). The folder should contain '.txt' scHi-C files in modified sparse upper triangular format (chr1, start1, chr2, start2, IF)}

\item{file.path.2}{Character string specifying the directory containing Hi-C data for the second condition (second cell-type group). The folder should contain '.txt' scHi-C files in modified sparse upper triangular format (chr1, start1, chr2, start2, IF)}

\item{imputation}{Character string or NULL of indicating the imputation method. Default is 'RF' for Random Forest imputation.}

\item{normalization}{Character string or NULL indicating the normalization method. Default is 'Loess'.}

\item{differential.detect}{Character string indicating the differential detection method. Default is 'MD.cluster'.}

\item{select.chromosome}{Integer or character indicating the chromosome to be analyzed (e.g., 'chr1' or 'chrX'.)}

\item{main.Distances}{Numeric vector or character 'full' indicating the interacting range (or full range) of genomic distances (in bp) for the method to focus.
Genomic distances (in bp) is the number of base pairs between two regions in the genome (e.g., loci or bins). Default is 1 to 10,000,000.}

\item{save.output.path}{Character string specifying the directory to save outputs, including the imputed cells in the form of a sparse upper triangular format,
normalization result table, and differential analysis result table. If NULL, no files are saved.}

\item{Plot}{A logical value indicating whether to plot the \code{differential.dect} results in an MD plot. Default is TRUE.}

\item{Plot.normalize}{A logical value indicating whether to plot the \code{normalization} results in an MD plot. Default is FALSE.}

\item{pool.style}{Character string specifying the pooling style for \code{imputation}. Options are 'progressive' or 'Fibonacci'. Default is 'progressive'. If the 'imputation
is skipped as NULL, the \code{pool.style} also should be NULL}

\item{n.imputation}{Integer specifying the number of imputations for the \code{imputation} step. Default is 5.}

\item{maxit}{Integer specifying the maximum number of iterations for internal refinement process within a single \code{imputation} cycle. Default is 1.}

\item{outlier.rm}{Logical. If TRUE, outliers are removed during \code{imputation}. Default is TRUE.}

\item{missPerc.threshold}{Numeric value specifying the maximum allowable percentage of missing data in each pool band.
Only pool bands within the \code{main.Distances} range and having a missing percentage below this threshold will be imputed
by the \code{imputation} method. Default is 95\%.}

\item{A.min}{A numeric value or NULL, specifying the A-value quantile cutoff to filter lower average expression in \code{differential.detect} step of \code{hic_compare()} function (from HiCcompare).
\code{hic_compare()} is used to detect outliers, which is assumed to be 'differences' bins in case of its number is too small (or none) to be cluster by GMM method.
If not provided, an optimized minimum A threshold that maximizes MCC and TPR while minimizing FPR in the simulated Hi-C matrix.}

\item{fprControl.logfc}{A numeric value controlling the false positive rate of \code{differential.detect} step by setting the threshold for the log fold change in the 'difference' cluster.
Detected differences identified by Gaussian Mixed Model (GMM) clusters only include values with log fold change that are larger than this threshold. Default is 0.8.}

\item{alpha}{A numeric value for the significance level of outlier detection of \code{differential.detect} step by the analysis of \code{hic_compare()} function from HiCcompare. Default is 0.05.}

\item{BP_param}{Parameters for \code{BiocParallel}, to be passed to the \code{bpparam()} function. See \code{?bpparam()} for more info.}
}
\value{
A list containing the differential analysis results and intermediate results (imputation, pseudo-bulk, normalization).
If \code{save.output.path} is provided , the imputed results for both conditions are saved in sparse format in the given firectory. Normalization and differential analysis results
are also saved if \code{save.output.path} is provided.
}
\description{
This function performs a differential analysis between two single-cell Hi-C data groups. It includes the
steps of imputation, normalization, and detection of differential chromatin interactions (DCIs).
}
\details{
This function implements the ScHiCcompare workflow. It first reads sparse Hi-C data from two conditions
and imputes missing interaction frequencies (if specified) using a random forest model (RF) along with a chosen
pooling method (either progressive or Fibonacci). In progressive pooling, distances are combined consecutively
to form larger sets, while Fibonacci pooling uses a Fibonacci sequence for combination.

Next, pseudo-bulk Hi-C matrices are generated, followed by joint normalization using Loess regression (from HiCcompare)
before detecting differential chromatin interactions via a Gaussian Mixture Model (GMM) clustering approach.

The differential analysis clusters normalized log fold changes in interaction frequencies between the two cell types
at each genomic distance into "difference" and "non-difference" groups. The non-difference group is assumed to
follow a normal distribution centered around 0 and is clustered using a Gaussian Mixture Model.
The difference cluster comprises points that belong to other distributions. If the size of the differences is
insufficient to form distinct distributions, these differences are considered outliers of the normal distribution,
identified by the \code{hic_compare()} function.
}
\examples{
\dontrun{
Load_example_MGFolder()
Load_example_ODCFolder()
result <- ScHiCcompare(
  file.path.1 = "MGs_example", 
  file.path.2 = "ODCs_example", 
  select.chromosome = "chr22", 
)
print(result)
}

}
