% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk_sparse.R
\name{pseudo_bulkHic}
\alias{pseudo_bulkHic}
\title{Generate Pseudo-bulk Hi-C Data}
\usage{
pseudo_bulkHic(scHiC.table, out = "sparse")
}
\arguments{
\item{scHiC.table}{A data frame containing interaction frequencies across single cells that is created by \code{scHiC_table} function.
The first four columns should represent 'cell', 'chr', 'region1', and 'region2',
followed by columns representing interaction frequencies ('IF') for individual cells.}

\item{out}{A character string specifying the output format. It must be either 'sparse' or 'full'.
Default is 'sparse'.}
}
\value{
A data frame representing the pseudo-bulk Hi-C data. If \code{out} is 'sparse',
it returns a data frame with pseudo-bulk matrix in format of a sparse upper triangular Hi-C matrix with five columns:
chromosome 1 (chr1), start position 1 (start1), chromosome 2 (chr2), start position 2 (start2), and interaction frequency (IF).  If \code{out} is 'full',
it returns a full matrix representation of the pseudo-bulk data.
}
\description{
This function generates pseudo-bulk Hi-C data from a single-cell Hi-C interaction frequency table.
It can return the data in either sparse or full matrix format.
}
\examples{
\dontrun{
# Load MG datasets folder example
load_example_MGfolder()
# Generate scHiC table for single cell Hi-C data of MG cell type
scHiC.table <- scHiC_table(file.path = "MGs_example",  cell.type = "MG", position.dataset = 1:50, type= "txt", select.chromosome = 22)

# Transform into the pseudo bulk HiC
pseudo_bulk_data <- pseudo_bulkHic(scHiC.table, out = 'sparse')
full_matrix_data <- pseudo_bulkHic(scHiC.table, out = 'full')
}

}
\references{
Stansfield JC, Cresswell KG, Vladimirov VI  et al (2018). Hiccompare: an R-package for joint normalization and comparison of HI-C datasets. BMC Bioinformatics  2018;19:279.
}
