% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHiC.table_MG_chr22.R
\docType{data}
\name{scHiC.table_MG_chr22}
\alias{scHiC.table_MG_chr22}
\title{scHi-C table from microglia (MG) cell type - chromosome 22 at 1 MB resolution}
\format{
A data frame with 335 rows and 5 columns:
\describe{
\item{cell}{Name of the target cell type}
\item{chr}{The chromosome of the interacting region's starting coordinate}
\item{start1}{The starting coordinate of the first interacting region}
\item{start2}{The starting coordinate of the second interacting region}
\item{IF_i}{The interaction frequency value for single cells, where i ranges from 1 to 10}
}
}
\source{
Single-cell data of MG cell type downloaded using \code{download_schic()} from \code{Bandnorm}.
See their website at \url{https://sshen82.github.io/BandNorm/articles/BandNorm-tutorial.html#download-existing-single-cell-hi-c-data}.
}
\usage{
data("scHiC.table_MG_chr22")
}
\value{
A data frame
}
\description{
A table that gathers 10 single-cell Hi-C data modified into a sparse upper triangular matrix.
It contains each cell's interacting region coordinates (cell, chr, start1, start2)
and the corresponding Interaction Frequencies (IF) for each single cell (IF_1, IF_2, ..., IF_10).
}
\examples{
data("scHiC.table_MG_chr22")
head(scHiC.table_MG_chr22)
}
\keyword{datasets}
